Fix definitions for clangd
In cpp it is pretty common to lookup definitions in a
file which is not already opened. So we make sure
to open.

Needed to slightly tweak the test helpers, because the compile
commands have absolute paths back to the original workspace
directory, so clangd will point back to the original file with the
definition.